+++ server/src/domain/use_cases/http_monitors/list_http_monitors_use_case.rs
@@ -15,7 +15,7 @@ use crate::domain::{
 #[ts(export)]
 pub struct ListHttpMonitorsResponse {
     pub items: Vec<HttpMonitor>,
-    pub total_number_of_results: u64,
+    pub total_number_of_results: u32,
 }
 
 #[derive(Error, Debug)]
